backend/requirements.txt
aiogram==3.13.1
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.2.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.115.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
makefun==1.15.1
passlib==1.7.4
pycparser==2.21
pydantic==2.9.2
PyJWT==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
starlette==0.38.6
typing_extensions==4.8.0
uvicorn==0.31.0
watchfiles==0.19.0
python-jose==3.3.0
passlib==1.7.4
requests==2.32.3
sqlalchemy==2.0.35
psycopg2==2.9.3
ipython==7.34.0
alembic==1.13.3
pydantic-settings==2.5.2
jinja2==3.1.4

# celry / redis
celery==5.4.0
redis==5.0.8
